requirements.txt
Flask==3.1.1
pysolr==3.10.0
psycopg2-binary==2.9.10
pandas==2.3.1